src/requirements.txt
# B-UEGMF Framework Dependencies
# 兼容Python 3.11的版本

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# NLP and Transformers
transformers>=4.20.0
tokenizers>=0.12.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Progress and Logging
tqdm>=4.65.0
colorlog>=6.7.0

# Development and Testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Cryptography (for blockchain simulation)
cryptography>=41.0.0

# System Monitoring
psutil>=5.9.0

# Configuration and Utilities
pyyaml>=6.0.0
click>=8.1.0
jupyter>=1.0.0
notebook>=6.5.0

# Development Dependencies (optional)
ipython>=8.0.0
ipdb>=0.13.0
memory-profiler>=0.61.0